pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomous-ai-software-factory"
version = "0.1.0"
description = "Autonomous multi-agent software factory with generation, analysis, repair, delivery, and observability."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy[asyncio]>=2.0.30",
  "asyncpg>=0.29.0",
  "aiosqlite>=0.20.0",
  "redis>=5.0.7",
  "celery>=5.4.0",
  "typer>=0.12.3",
  "PyJWT>=2.8.0",
  "python-multipart>=0.0.9",
  "email-validator>=2.2.0",
  "httpx>=0.27.0",
  "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
]

[project.scripts]
factory = "cli.main:main"

[tool.setuptools]
packages = [
  "agents",
  "api",
  "cli",
  "core",
  "domain",
  "infrastructure",
  "orchestration",
  "plugins",
  "services",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --cov=agents --cov=core --cov=domain --cov=services --cov=orchestration --cov=plugins --cov-report=term-missing --cov-fail-under=90"

[tool.coverage.run]
branch = true
omit = [
  "core/config.py",
  "domain/ports.py",
]

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = [".venv", "workspaces", "external_plugins"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC", "SIM"]
ignore = ["B008", "ASYNC240", "E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unreachable = true
plugins = []
